requirements.txt
duckdb
pandas
openpyxl
openai
python-dotenv
